Update to 0.38.3: Fixed bug introduced with 0.38 where the preview for some results would show as a thin column

+++ XUI/AlbumArtDownloader/Controls/ArtPanel.cs
@@ -495,7 +495,7 @@ namespace AlbumArtDownloader.Controls
 			double imageAspectRatio = 1; //If no album art yet, then assume it will be square.
 			if (AlbumArt != null)
 			{
-				imageAspectRatio = AlbumArt.ImageWidth / AlbumArt.ImageHeight;
+				imageAspectRatio = (double)AlbumArt.ImageWidth / (double)AlbumArt.ImageHeight;
 			}
 			
 			//Popup may not take up more than 75% screen area (matches WPF Popup restriction)
